changes for using with tkeycoin-core-2

+++ src/serialization/UrlSerializer.cpp
@@ -39,8 +39,8 @@ struct tokens: std::ctype<char>
         static cctype::mask rc[cctype::table_size];
         std::memcpy(rc, const_rc, cctype::table_size * sizeof(cctype::mask));
 
-        rc['&'] = std::ctype_base::space;
-        rc[' '] = std::ctype_base::print;
+        rc[static_cast<size_t>('&')] = std::ctype_base::space;
+        rc[static_cast<size_t>(' ')] = std::ctype_base::print;
         return &rc[0];
     }
 };
